Organize API tests into domains
